Fixed Simon Fau`s case.

+++ pyibisami/ami_model.py
@@ -342,8 +342,8 @@ class AMIModel:
 
         if bits_per_call:
             self._bits_per_call = int(bits_per_call)
-        bits_per_call = self._bits_per_call
-        samps_per_call = self._samps_per_bit * bits_per_call
+        bits_per_call = int(self._bits_per_call)
+        samps_per_call = int(self._samps_per_bit * bits_per_call)
 
         # Create the required C types.
         Signal = c_double * samps_per_call
